- Adapt auditing for deleted posts and make it more efficient. - Adapt UI so it does not tries to show deleted posts. - Add delete button for posts when a user visits their own profile. - Rename React components for more clarity.
